added default message to annotation + renamed package to plural
