+++ api/src/main/java/tennisclub/annotations/IsEndTimeAfterStartTime.java
@@ -1,9 +1,8 @@
 package tennisclub.annotations;
 
-import tennisclub.validator.IsEndTimeLaterThanStartTimeValidator;
+import tennisclub.validators.IsEndTimeLaterThanStartTimeValidator;
 
 import javax.validation.Constraint;
-import javax.validation.Payload;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -16,6 +15,8 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Constraint(validatedBy = IsEndTimeLaterThanStartTimeValidator.class)
 @Documented
 public @interface IsEndTimeAfterStartTime {
+    String message() default "End date must be greater or equal to start date.";
+
     String start();
 
     String end();
